requirements.txt
urllib3==1.25.8
allure-pytest==2.8.13
pylint==2.4.4
pytest==5.3.5
selenium==3.141.0
pytest-xdist==1.31.0
requests==2.23.0
sshtunnel==0.1.5
mysql-connector-python==8.0.19
paramiko==2.7.1